finalize groups fully crud

+++ app/Http/Controllers/Stock/MainGroupController.php
@@ -8,7 +8,7 @@ use App\Models\Stock\StockGroup;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\Stock\StoreResource;
 use Symfony\Component\HttpFoundation\Response;
-use App\Http\Resources\Stock\MainGroupResource;
+use App\Http\Resources\Stock\StockGroupResource;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use App\Http\Requests\Stock\MainGroups\StoreMainGroupRequest;
 use App\Http\Requests\Stock\MainGroups\UpdateMainGroupRequest;
@@ -23,19 +23,19 @@ class MainGroupController extends Controller
     public function index(): View
     {
         $lastGroupNr = StockGroup::isRoot()->latest()->first()?->id + 1 ?? 1;
-        $groups = StockGroup::isRoot()->get();
+        $groups = StockGroup::isRoot()->paginate(5);
         return view('stock.MainGroups.index', compact('lastGroupNr', 'groups'));
     }
 
     /**
      * store
      * @param StoreMainGroupRequest $request
-     * @return MainGroupResource
+     * @return StockGroupResource
      */
     public function store(
         StoreMainGroupRequest $request
-    ): MainGroupResource {
-        return MainGroupResource::make(
+    ): StockGroupResource {
+        return StockGroupResource::make(
             StockGroup::create($request->validated())
         )
             ->additional([
@@ -47,12 +47,12 @@ class MainGroupController extends Controller
     /**
      * show
      * @param  StockGroup $stockGroup
-     * @return MainGroupResource
+     * @return StockGroupResource
      */
     public function show(
         StockGroup $stockGroup
-    ): MainGroupResource {
-        return MainGroupResource::make($stockGroup)
+    ): StockGroupResource {
+        return StockGroupResource::make($stockGroup)
             ->additional([
                 'message' => null,
                 'status' => Response::HTTP_OK
@@ -69,9 +69,9 @@ class MainGroupController extends Controller
     public function update(
         UpdateMainGroupRequest $request,
         StockGroup $stockGroup
-    ): MainGroupResource {
+    ): StockGroupResource {
         $stockGroup->update($request->validated());
-        return MainGroupResource::make($stockGroup)
+        return StockGroupResource::make($stockGroup)
             ->additional([
                 'message' => "تم تعديل المجموعةالرئيسية بنجاح",
                 'status' => Response::HTTP_OK
@@ -86,6 +86,12 @@ class MainGroupController extends Controller
     public function destroy(
         StockGroup $stockGroup
     ): JsonResponse {
+        if ($stockGroup->hasChildren()) {
+            return response()->json([
+                'message' => 'لايمكن حذف المجموعة لانها تحتوى على مجموعات فرعية',
+                'status' => Response::HTTP_UNPROCESSABLE_ENTITY
+            ]);
+        }
         if ($stockGroup->delete()) {
             return response()->json([
                 'message' => 'تم حذف المجموعةالرئيسية بنجاح',
